+++ src/app/lib/lforms-widget-registry.ts
@@ -73,7 +73,10 @@ import {
   CalculatedInitialExpressionComponent
 } from "./widgets/expression-editor/calculated-initial-expression/calculated-initial-expression.component";
 import {EnableWhenExpressionComponent} from "./widgets/expression-editor/enable-when-expression/enable-when-expression.component";
+import {ExtensionPrimitiveComponent} from "./widgets/extension-primitive/extension-primitive.component";
+import {MinMaxOccursComponent} from './widgets/min-max-occurs/min-max-occurs.component';
 import {ExtensionUrlComponent} from './widgets/extension-url/extension-url.component';
+import {QuestionnaireHiddenComponent} from './widgets/questionnaire-hidden/questionnaire-hidden.component';
 
 @Injectable()
 export class LformsWidgetRegistry extends DefaultWidgetRegistry {
@@ -132,6 +135,7 @@ export class LformsWidgetRegistry extends DefaultWidgetRegistry {
     this.register('variable', VariableComponent);
     this.register('initial-number', InitialNumberComponent);
     this.register('entry-format', EntryFormatComponent);
+    this.register('questionnaire-hidden', QuestionnaireHiddenComponent);
     this.register('answer-value-set-coding-display', AnswerValueSetCodingDisplayComponent);
     this.register('extension', ExtensionComponent);
     this.register('extension-container', ExtensionContainerComponent);
@@ -143,6 +147,8 @@ export class LformsWidgetRegistry extends DefaultWidgetRegistry {
     this.register('coding-display', CodingDisplayComponent);
     this.register('enable-when-method', EnableWhenMethodComponent);
     this.register('enable-when-expression', EnableWhenExpressionComponent);
+    this.register('extension-primitive', ExtensionPrimitiveComponent);
+    this.register('min-max-occurs', MinMaxOccursComponent);
     this.register('meta-profile', MetaProfileComponent);
     this.register('usage-context', UsageContextComponent);
     this.register('usage-context-code', UsageContextCodeComponent);